The stateful image URL should not contain "/update" in it
When image_to_live is run with a URL, this URL is placed directly into CHROMEOS_DEVSERVER, including the trailing "/update".  This works perfectly for all of the rest of image_to_live, and worked with older versions of stateful_update.  However with the recent changes, instead of using omaha_ping, stateful_update reaps the URL out of /etc/lsb-release with the "/update" in it.

This change filters "/update" out of the URL.

TEST=Rerun image_to_live

Review URL: http://codereview.chromium.org/3163026