Reworked devserver so that update images generated are cached in directories named after the md5 hashes of the images used to generate the updates. Only the 12 most recent deltas are preserved. All deltas generated (outside of the factory codepath) are modified.
The purpose of this is to speed up delta generation during automated tests, especially when
running autotest and/or ctest.

autoupdate.py has been refactored in a number of small ways.

All of the Generate methods now return the name of the update generated (or retrieved from
cache), instead of the update file being assumed to be static/update.gz.

Update files are now created directly in the cache directories instead of in the image
directory and then copied.

--clear_cache has been added to devserver.py to force clearing of all cached update values.

--use_cached no longer does anything, but has not yet been removed. It appears to be
intended to force the use of static/update.gz, but I can't find any cases where it's used.

TEST=The autoupdate unit test has been udpated and is passing.
./image_to_live.sh has been tested by hand with both full image and delta updates.
./ctest can run successfully.

BUG=chromium-os:8207  (early prep)

Review URL: http://codereview.chromium.org/4906001

Change-Id: I19125f0668a4f70cd569ce7aa992aa583daeaf30