Add --payload to devserver to allow the update payload to be specified by the devserver user.
Should have no effect if --payload is not specified, but should cause update.gz and stateful.tgz to be symlinked to
the specified payload files if it is.

Right now, it expects you to specify the file update.gz, and for stateful.tgz to be in the same directory. That could easily be changed to only specify the directory.

BUG=chromium-os:8207
TEST=Ran by hand.

Review URL: http://codereview.chromium.org/5109004

Change-Id: I586a64119a44e85033269014e6e1d8932e586b7a