Add support for --proxy_port.
For some test programs, we want incoming update requests to come to an
alternate port from the devserver port, and to run the dev server at the
normal port (there will be a proxy between the two that manipulates the
data).

This new option directs the client to the proxies port without changing
the server port.

Change-Id: If4bd7a9681bdafd8412c2a0200fd9bf364f501ef

BUG=chromium-os:8207
TEST=

Review URL: http://codereview.chromium.org/5611004