Temporarily disable flash parameter configuration.
The flash parameters need to come after the BCT file in the
cbootimage config file (otherwise they are clobbered by the
BCT's values).  But once they are moved to the correct location
the resulting image doesn't boot.  This change means that
we can only boot from flash devices that are already specified
in the BCT file.

BUG=chromium-os:11981
TEST=Build and Boot from spi on Kaen.

Change-Id: Ic5e92e8057c56bb5c5e2f0e3ca191806c56859f0

Review URL: http://codereview.chromium.org/6609004