Add --dry_run flag to cros_write_firmware.
This flag causes cros_write_firmware to do all steps except for
calling nvflash.  This is useful for verifying that all required
components and tools are available.  It is also useful in
conjunction with the --config option to save a signed firmware
image for later use.

BUG=None
TEST=run cros_write_fimware with and without --dry_run.  Verify that
without the flag the image is written correctly and that with the
flag nvflash is not run and thus the device does not get updated.
Then run with --dry_run and --config save and use flashrom to write
save/signed.bin to SPI on the target and reboot, verifying that the
signed.bin file is a valid firmware for Tegra.

Change-Id: I328273fb33fa37feaed13f5e55fc8f673f227f5e

Review URL: http://codereview.chromium.org/6713134