Make willis show relative paths.
The "what's going on" tool displays the repositories with
modifications, branches, etc., prepending each repo with a
path starting at ~/trunk.

It would be more convenient if it was showing the path
related to the current directory - this way the path can be
used directly as argument to pushd/sd/ls/etc.

This CL modifies the output to show the relative path and
removes some redundant code.

Change-Id: Ib25e26f27778a1c5b8d7d3e611deeba60294d1af

BUG=None
TEST=manual:

Run `willis' from different directories in chroot. Observe
the results reporting relative paths to the appropriate git
repositories.

Review URL: http://codereview.chromium.org/6824011