Willis: don't skip repositories with branches that are headless.
Currently willis will ignore any repository that is in the
headless state.  This was initially an attempt to just ignore
repositories that had no branches.  But the check is too broad.
If you have made a branch and then checked out m/master your
repository should still show up in willis' list.  The solution
is to only ignore repositories that are headless and have no other
branches.

BUG=None
TEST=run willis, verify that all repositoris with branches are shown.

Change-Id: Idc21dd8eefadb8941922d14a69cc17dd82a3f47d

Review URL: http://codereview.chromium.org/6894010