Update devserver to filter packages in a separate root.
The devserver wants to clean debug symbols and build-time only files
from packages before installing to the device, but doing this directly
in the build root destroys our build environment. To prevent this
destruction, I've moved the filtering to occur in a subdirectory.

This task requires some extra copying, as we need to copy the files to
a separate root before editing them, but it makes it possible to easily
gmerge huge packages like chromeos-chrome to a small root partition.

BUG=chromium-os:12677
TEST=gmerge chromeos-chrome and a few other packages to a VM.

Change-Id: I085025d8ea908bf1801beece938d08064bf1ad2e
Reviewed-on: http://gerrit.chromium.org/gerrit/1138
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: David James <davidjames@chromium.org>