Revert "Teach Portage to grab portage config from stateful partition."
This reverts commit 1f09a107cea3d5aacf51572cde12c56f1251141e.
Description below.

Teach Portage to grab portage config from stateful partition.

With PORTAGE_CONFIGROOT=/usr/local, Portage looks in /usr/local/etc for config
instead of /etc. This allows us to configure necessary configstuffs in
the stateful partition and avoids the need to copy it over to the rootfs
manually.

I've also updated ROOT to only be a default so that if the user can set
ROOT=/usr/local from the commandline if they want.

BUG=chromium-os:12474
TEST=Verify that gmerge works with this change, and that
     gmerge -n mesa works now (it previously didn't work because
     it depended on packages in package.provided).

Change-Id: I9f88fc87547578a8deff06d458274ed540a901d7
Reviewed-on: http://gerrit.chromium.org/gerrit/1214
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: David James <davidjames@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/1232
Reviewed-by: David James <davidjames@chromium.org>