Add repo-related bash completion.
This is heavily based on the script that came from the ChormiumOS
sites page.  Currently, it is here:
   http://www.chromium.org/tips-and-tricks-for-chromium-os-developers
That script will be removed once this one is checked in (and a reference
will be made to this file in git).

This file adds some functions that are useful for using in your PS1,
much like the standard git bash completion file.

A future CL will install this in the cros-devutils ebuild.

BUG=chromium-os:13186
TEST=In combination with cros-devutils ebuild, saw that __git_branch_ps1
was there and worked.

Change-Id: I7681077abc27ef63ed19a070a2638d3a3c8659da
Reviewed-on: http://gerrit.chromium.org/gerrit/1574
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Darin Petkov <petkov@chromium.org>