Create developer mode install script
This directory will contain all the necessary files to add
some capabilities to a base image booted in developer mode. It
will download and bootstrap first a version of emerge so the
user can install packages. It will also give the option of
installing a set of developer packages.

BUG=chromium-os:16841
TEST=None till actual code is added. This is a first commit
adding the needed files.

Change-Id: I17aae26a164ab361de059840749a5977f2ec3947
Reviewed-on: http://gerrit.chromium.org/gerrit/3122
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Arkaitz Ruiz Alvarez <arkaitzr@chromium.org>