Added ability to install missing compilers when using cros_gcc_config.
Example: cros_gcc_config i686-pc-linux-gnu-4.6.0

will install cross-i686-pc-linux-gnu/gcc-4.6.0 and copy over the
necessary libraries to the relevant boards.

Also fixed a bug with copying over libs for arm.

BUG=none
TEST=cros_gcc_config i686-pc-linux-gnu-4.6.0
This installed gcc-4.6.0 with the correct use flags even if gold was the
default linker.
cros_gcc_config <number> to switch to the config of that number.

Change-Id: I40ce020bdcc9513df172a937d177b85208e37aff
Reviewed-on: http://gerrit.chromium.org/gerrit/3532
Tested-by: asharif <asharif@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>