Have cros_write_firmware not reference /u-boot/u-boot.dtb
The u-boot.dtb file is installed by the u-boot build right now and
is different for every board.  Since we want the u-boot build to be
binary-compatible all tegra boards, we need to stop installing that
file in the u-boot ebuild.

In cros_write_firmware, we'll now default to using one of the dtb
files based on the board name.

BUG=chromium-os:17191
TEST=Tried running cros_write_firmware w/ no u-boot.dtb.

Change-Id: Id680a05ccd53fd5379a999f362ae469251f323a0
Reviewed-on: http://gerrit.chromium.org/gerrit/4265
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Anton Staaf <robotboy@chromium.org>