Add make.defaults, start coding dev_install.sh
Since this package should not interfere with the portage configuration in the
chroot we use make.defaults and install it to /build/$BOARD/etc/portage/. The
target image will not include make.conf, so it will use make.defaults.
dev_install.sh also includes code to download a bz2 file to bootstrap emerge.
Data repositories need still to be configured.

BUG=chromium-os:6943
TEST=Check the rootfs in the target for /etc/make.defaults and the
absence of /etc/make.conf and /etc/portage/make.conf

Change-Id: I365e81c7c6bd3548f566232b51e512f8bcebd12c
Reviewed-on: http://gerrit.chromium.org/gerrit/4060
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Arkaitz Ruiz Alvarez <arkaitzr@chromium.org>