Add postload support to cros_bundle_firmware
A postload binary file contains code that is not part of the main U-Boot
image. It can be loaded from flash later when needed.

This commit adds support for this extra file, which is placed in the output
file immediately after the signed bootstub.

BUG=chromium-os:17753
TEST=~/trunk/src/platform/dev/host/cros_bundle_firmware -w -u u-boot.bin -O out -v2 -P u-boot-post.bin
Check that resulting image boots, then:
postload recovery
ext2ls mmc 0:5
(this executes even though the ext2 code is in the postload)

Change-Id: I4b4b5b814ea2a8f420382af76ee57e67699ef9ba
Reviewed-on: http://gerrit.chromium.org/gerrit/4904
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>