Update cros_write_firmware to use the new find_u_boot_component function
This change updates cros_write_firmware to use the new find_u_boot_component
function so that it can find its inputs in either the existing "u-boot"
directory or the coming "firmware" directory.

BUG=chromium-os:18107
TEST=Modified some u-boot related ebuilds to put some files in the firmware
directory. Ran the ebuilds and verified that the files showed up where
expected. Ran cros_write_firmware and verified it found the files in either
location.

Change-Id: I35d81bfa6d465bd4369fd7656781c43fd360564c
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/5175
Reviewed-by: Doug Anderson <dianders@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>