cros_bundle_firmware: Don't use --secureboot with the flasher
The flasher needs to use the normal boot mechanism since it sets up quite a long
bootcmd. This change keeps the original unchanged fdt around for use with the
flasher, so that it will still work when using secure boot.

The change is made in cros_write_firmware also just to keep it consistent.

BUG=chromium-os:17298
TEST=~/trunk/src/platform/dev/host/cros_bundle_firmware -w -u u-boot.bin -s -v2 -b tegra2_aebl --bootsecure --bootcmd vboot_twostop
See that the flasher still works

Change-Id: I5b1b1b9c7e96b662b4a969597bf9ef20733e8804
Reviewed-on: http://gerrit.chromium.org/gerrit/5594
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>