Introduce Dev Server API. Adds hostinfo, setnextupdate methods.
Adds a couple methods in preparation for AU testing where no SSH access
to the host is available. Instead information about the update process
must be gleaned from the queries made to the Dev Server.

hostinfo provides information on the last received update type and
status as well as known version information. Information is returned in
a JSON dictionary.

setnextupdate takes an update label (as normally provided to /update)
and causes the next update ping from that host to use that url for the
update.

BUG=none
TEST=Ran Dev Server. Queried API w/ valid and invalid. Verified next
update used set one.
http://172.31.26.251:8081/api/hostinfo/172.31.26.144

Change-Id: I907b419371aab5774c618fa9192ef056cd6aa046
Reviewed-on: http://gerrit.chromium.org/gerrit/5933
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Dale Curtis <dalecurtis@chromium.org>