cros_workon - temporarily special case chrome package
When the user cros_workon starts the chrome package for a board, we set
the CHROME_ORIGIN envvar to GERRIT_SOURCE, as part of the new chrome
development workflow.  The envvar is set in the
${board}/etc/portage/env/chromeos-base/chromeos-chrome file.  We delete
the file when the user runs cros_workon stop for the chrome package.

This change is temporary - GERRIT_SOURCE will eventually become the
default behavior.

BUG=chromium-os:19112
TEST=Verified ebuild behavior for the cros_workon start and stop cases.

Change-Id: I4948383bac3fc4eafc211da2058cbcb24614cedc
Reviewed-on: http://gerrit.chromium.org/gerrit/6049
Reviewed-by: Ryan Cui <rcui@chromium.org>
Tested-by: Ryan Cui <rcui@chromium.org>