cros_bundle_firmware: Use default bmpblk file
We have moved to pre-built bmpblk files and it is no longer possible for
cros_bundle_firmware to build one since the required script has been removed.
It is convenient for devs to use the default bmpblk rather than specifying
it each time.

BUG=chromium-os:19724
TEST=manual: cros_bundle_firmware
See that it runs correctly and uses default bmpblk

Change-Id: Ia04c5dc344eee0a2932f500beb06961298a7bc1b
Reviewed-on: http://gerrit.chromium.org/gerrit/6763
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>