Use the new "bios-base" option to adjust the offset of sections in the fmap
When generating the binary fmap to include in the image, use the bios-base
setting in the "flash" node of the device tree to adjust the section offsets
so they reflect offsets from the start of the ROM. That should allow tools like
flashrom to work with them without any special handling.

BUG=chrome-os-partner:5844
TEST=Used this and other changes to compensate for the image offset and vboot
on stumpy. Vbooted on x86-alex. With some help from dhendrix, used the
fmap_decode utility to check that the binary fmap was adjusted at the right
time and by the appropriate amount.

Change-Id: I24a26751ff10bdac70ce4db9ed1db9b068e0ef95
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/7241
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>