Have cros_chrome_make run hooks when user-specified USE flags change.
USE flags can change the GYP_DEFINES env var, which changes the
Makefiles that would be generated by gyp_chromium, so we need to
regenerate the Makefiles when user-specified USE flags change.

Also stops looking at temporary files to tell if previous run was
GERRIT_SOURCE or LOCAL_SOURCE.  Instead, check the symlink to
CHROME_ROOT in the working directory.  This is less prone to error/bugs.

BUG=chromium-os:19903
TEST=Checked that a change in USE env var forces runhooks.
     Checked that switching --build_tests force runhooks.
     Checked that no change does not force runhooks.
     Checked switching between LOCAL_SOURCE and GERRIT_SOURCE forces
runhooks.
     Checked doing an 'ebuild clean' forces runhooks

Change-Id: I15850579eba88de6fef8862b8226ccb045209987
Reviewed-on: http://gerrit.chromium.org/gerrit/7754
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Ryan Cui <rcui@chromium.org>