Add some very basic NAND support to cros_write_firmware.
NOTE that NAND is not really well-tested or well supported
still, but this at least gets us something that seems to work.

This uses bct_dump to tell whether we want NAND or SPI.
The BCT that is looked at is the one from the _flasher_,
not anything that might be part of the firmware that is
being written.  Hopefully that's OK.

This is part of a batch of CLs:
- http://gerrit.chromium.org/gerrit/7838
  Hack it so that bundle_firmware forces one copy of the BCT for NAND.
- http://gerrit.chromium.org/gerrit/5207
  CHROMIUM: cbootimage: Add Bctcopy parameter
- http://gerrit.chromium.org/gerrit/7841
  Revert "Hack tegra2_dev-board to think NAND block size is 64K."
- http://gerrit.chromium.org/gerrit/7837
  Add some very basic NAND support to cros_write_firmware.
- http://gerrit.chromium.org/gerrit/7840
  CHROMIUM: harmony: Very basic harmony (AKA dev-board) support.

BUG=chromium-os:17465
TEST=Used cros_write_firmware on both SPI Flash (tegra2_kaen) and
NAND (tegra2_dev-board).

Change-Id: I318ba512824317a313923009443902601d1fe26e
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/7837