Make willis a bit more lively.
Even after being parallelized willis still takes a while to complete
in some cases, leaving the operator at loss as of what's happening.

This modification adds a single line status update which changes each
time a thread starts or finishes, reporting something like

  'Started threads: 238, remaining threads:  19'

showing countdown of running threads to zero.

Also, a command line option -f/--fast_report is being added. If this
option is present, the repository query results get reported
immediately (which will cause them to be printed out of order and
likely in different order on different runs). If this option is set,
the status update is not printed.

BUG=none
TEST=manual

  . ran willis multiple times with and without the -f option. Observed
    expected behavior (status updates when -f is not present, fast
    reporting when -f is present).

Change-Id: I812b6b6adf88b2d4117e011e26c3e712393a2043
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/10851
Reviewed-by: Anton Staaf <robotboy@chromium.org>