cros_bundle/write_firmware: Check U-Boot text base in image
This is only meaningful on ARM.

Normally text base is in the fdt, so we just read it from there. But as
a second check we look at the image itself in case this is different, and
switch to that if it is.

We will move over to a new address on the next rebase, but this keeps
us running for now.

BUG=chromium-os:19724
TEST=emerge chromeos-bootimage on Seaboard and Alex

Change-Id: I322169f0dd6054a9decd1bac5dde09301d805285
Reviewed-on: https://gerrit.chromium.org/gerrit/11744
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>