Teach gmerge to blow away cache in between runs.
If Packages file is hand-edited in /build/board/packages, gmerge should
pick up the change right away. Right now the change is only picked up if
the TIMESTAMP entry in the file changes, which might not happen if the
file is edited by hand.

BUG=chromium-os:23113
TEST=Verify gmerge -n chromeos-chrome takes into account hand-edits to
     the packages file on the server.

Change-Id: Id0770ab06658fe31c2c63738f4e27da8966e3b77
Reviewed-on: https://gerrit.chromium.org/gerrit/11833
Reviewed-by: Nirnimesh <nirnimesh@chromium.org>
Tested-by: David James <davidjames@chromium.org>
Commit-Ready: David James <davidjames@chromium.org>