Add logic of whether or not we patched the kernel to the devserver cache.
The current devserver caching logic assumes payloads are the same if they have
the same src_image, target_image and private key. For callers that build both
vm and non-vm payloads i.e. with/without a patched kernel, these caching
assumptions are no longer true as patching the kernel affects the payloads.
This change factors in whether or not patch the kernel in the payload name
so that we don't cache the same payloads with/without a patched kernel onto
the same payload.

BUG=chromium-os:25622
TEST=Ran with/without --for_vm on my latest image with a full payload saw
them cached in different locations.

Change-Id: Ie5f2309e0a79f69a0181608290dc1b10d225d86b
Reviewed-on: https://gerrit.chromium.org/gerrit/15103
Commit-Ready: Chris Sosa <sosa@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>