Update devserver logging to utilize --log-dir.
Since not everyone wants to log files and depending on if the devserver
is run inside or outside of the chroot it may cause problems, if we assume
we can write in the directory where dev server is installed add a --log-dir
option. If specified dump access and error logs in to that directory.

TEST=Ran with --log-dir and a valid dir to ensure files were logged,
Ran --log-dir without a valid dir to ensure error.
Ran without --log-dir to ensure no log files were created.
BUG=chromium-os:26870

Change-Id: I2ebd946be61e4caf33cdb0769e262cee59eb85f1
Reviewed-on: https://gerrit.chromium.org/gerrit/16815
Reviewed-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Mark Hayter <mdhayter@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>