cros_bundle_firmware: update fdt.py for upstream fdtget/put
These tools have now been upstreamed, so it is time to move this library
over to the upstream version. This updates the interface and the unit
tests.

It also updates those tools which use it to use the new interface.

BUG=chromium-os:23249
TEST=run unit tests:
$ cd .../src/platform/dev/host/lib
$ python lib/fdt.py --test
See that all tests pass

Change-Id: I5b203278d2381cf80299aa3a6c96fbe5a3fa03d9
Note: This change should be squashed into the previous one.
Reviewed-on: https://gerrit.chromium.org/gerrit/17525
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>