cros_bundle_firmware: Refactor packer blob code
At present we have hard-coded members for the various files that we can
place in blobs. Move to dictionary approach, which is more extensible.

pack_firmware.GetBlobList() looks through the flash map and returns a
list of blob types that we are going to need to produce. We then produce
each one in term, passing it to pack_firmware via AddProperty().

BUG=chromium-os:19724
TEST=manual: run cros_bundle_firmware
Load resulting image onto the board and see that it boots ok

Change-Id: I1051f818e724276d7e9f757d811a8696bd18502b
Reviewed-on: https://gerrit.chromium.org/gerrit/17770
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>