cros_bundle_firmware: Add support for Exynos BL1/BL2 entries
The exynos SOCs require a BL1 pre-boot and a BL2 SPL boot to be stored
in flash before U-Boot.

For BL1 we just use a provided binary. This is signed by Samsung and is
designed to cope with chip-specific initialization.

For BL2 we use U-Boot's SPL (Secondary Program Loader). This is
responsible for setting up SDRAM and loading U-Boot.

BUG=chromium-os:19724
TEST=manual:
PYTHONPATH="~/cosarm/src/platform/dev/host/lib" python
 ~/cosarm/src/platform/dev/host/cros_bundle_firmware -s -b daisy -I
 ~/u/board/samsung/dts --dt exynos5250-smdk5250.dts -I ~/u/arch/arm/dts
 -u ~/u/u-boot.bin --bl1 ~/cosarm/chroot/build/daisy/firmware/E5250.nbl1.bin
 --bl2 ~/u/spl/u-boot-spl.bin -O out
See that we get correct output in the 'out' directory

Change-Id: I5625a81744709608ae8cb3edf39115edb85b5ba3
Reviewed-on: https://gerrit.chromium.org/gerrit/18052
Commit-Ready: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>