cros_bundle_firmware: Hack to support Tegra during transition
For now, Tegra Flashmap files do not include a 'required' property, so
using the -s flag will produce an empty image. If we see that there are
no required entries, then we select the first 'signed' blob that we see.
This corresponds to the signed U-Boot image on Tegra.

This hack will be removed when Tegra .dtb files are updated and back in
our tree. For now it allows us to flash Tegra boards.

BUG=chromium-os:19724
TEST=manual: run cros_bundle_firmware -w
See that it flashes an image to the Seaboard and that it boots

Change-Id: Ifaa2f15c2bdf23dc533019c2089d1a7e73f64a90
Reviewed-on: https://gerrit.chromium.org/gerrit/18323
Commit-Ready: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>