Always send ChromeOSVersion in Omaha response.
With the changes in update_engine and Omaha for enterprise control for
auto-update behavior, Omaha could potentially hand over a older version
of the OS to a client that already is on a higher version. So, we will be
adding support in update_engine to ignore updates that don't have a
ChromeOSVersion field or updates whose ChromeOSVersion is less than the
current version. So, in order for dev server to continue to work, we'd
have to add the ChromeOSVersion attribute in Omaha response. The actual
version number should be higher than the current version. Since the
current version number is not available in this context, I'm hardcoding
9999.0.0 as the version which will make sure update_engine will always
honor any updates from devserver as it does today.

BUG=27308:Support enterprise auto-update only to a given targetversion
TEST=Tested on ZGB with new devserver and made sure it updates correctly.
CQ-DEPEND=I61a498280c3ccca674e74581f1b03364d7896bc2
Change-Id: I523c3f67e0cb07fd24744dc0a30382ff2fe2128a
Reviewed-on: https://gerrit.chromium.org/gerrit/18417
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Jay Srinivasan <jaysri@chromium.org>
Tested-by: Jay Srinivasan <jaysri@chromium.org>
Commit-Ready: Jay Srinivasan <jaysri@chromium.org>