cros_output: Move progress clearing to correct place
If we clear progress information on any message, even one that is not
actually going to be sent to output, then we end up losing a lot of
progress information. It flashes on the screen quickly and then goes
away.

Fix this problem by only clearing progress when a message is actually
being sent.

BUG=chromium-os:19724
TEST=write cros_bundle_firmware; see that progress messages now persist

Change-Id: I8242887ad1531b4e07482577aaef592719e5352b
Reviewed-on: https://gerrit.chromium.org/gerrit/18865
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>