cros_bundle_firmware: Use servo to control board flashing
OK, I give up. Flashing the board by manually holding buttons is just too
painful. We have to cold reset the board, then hold fw-up and the power
button while starting the flash with the other hand, and then keep the
power button down until it has finished.

It seems that we are going to have to require servo for flashing Daisy for
now, so add this to cros_bundle_firmware.

BUG=chromium-os:19724
TEST=write cros_bundle_firmware; see that it controls servo and flashes
successfully

Change-Id: I61c8baeb31d99cd19abf1a4333bdb2870f6d771b
Reviewed-on: https://gerrit.chromium.org/gerrit/18867
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>