cros_bundle_firmware: Tidy up exynos bl2 handling
There was a bug where the nul terminator of the machine parameter
string was not counted, so fix that.

Also tidy up the progress and logging output for the user. Use the tools
ReadFile/WriteFile instead of copyfile so that file names and sizes are
logged.

BUG=chromium-os:28242
TEST=cros_bundle_firmware -v3; see that messages appear for progress and
also file size.

Change-Id: I4687a4a48a9302b62cf30cbf3eb2e306ec558a48
Reviewed-on: https://gerrit.chromium.org/gerrit/18931
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>