cros_bundle_firmware: Support -F to specify the flash destination
The -F flag indicates that we want to create a flasher program which will
write the image to flash on the device. This flasher program can then be
written to the target as specified by the -w option.

The argument to -F is the flash memory that will be written by the image.
Valid options are sdmmc, spi and nand, depending on target support.

BUG=chromium-os:28242
TEST=use cros_bundle_firmware with and without -F to flash seaboard
and daisy.

Change-Id: I1d234faacb9ce900ab234965ae16364381d3f3f8
Reviewed-on: https://gerrit.chromium.org/gerrit/18933
Commit-Ready: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>