cros_bundle_firmware: Try to find a default FDT file
Where no FDT is specified it is nice to try pretty hard to find the
correct one. So use a wildcard to pick up something based on the board
name. Provided that we only find one match this should be safe enough.

BUG=chromium-os:28958
TEST=manual:
$ cros_write_firmware -b daisy -i /build/daisy/firmware/legacy_image.bin -w sd:
See that it automatically finds the FDT file now.

Change-Id: I76187cf48c1428b2056372cf02ae6ffce8f1e1ed
Reviewed-on: https://gerrit.chromium.org/gerrit/19597
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>