Add --noreboot_after_update to image_to_live.sh.
This is an image_to_live.sh that is only useful if you are debugging
update related stuff. But it's useful to me, since it leaves the target in
an 'update installed but not live yet' state. If I knew how to move
the help message into an 'advanced' section, I would.

BUG=None
TEST=Used it.

Change-Id: I72ebee69dfaf76ea2f0ab13e78c5c476499e8d58
Reviewed-on: https://gerrit.chromium.org/gerrit/19784
Commit-Ready: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>