cros_bundle_firmware: Move some tool checks to write_firmware
Rather than checking for the flashing tools right at the start, do this
when we know we are going to need to write, and which particular board
we are writing to.

BUG=chromium-os:28958
TEST=manual:
Remove /usr/bin/nvflash
$ cros_bundle_firmware -b daisy -w sd:.
See that we get no warning now.

$ cros_bundle_firmware -w usb
See that we do get a warning.

Change-Id: Ia99580dff11a0fea55d102d0010c3e30153f9303
Reviewed-on: https://gerrit.chromium.org/gerrit/19843
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>