[dev-util] Remove server.socket_timeout from dev server
This is what's causing us to wait for ages when we post crash dumps
to the dev server for symbolication.  It was put there as a part of
an attempt to fix an issue gmerg'ing chrome, but we're pretty sure
that other timeouts added as a part of the same change are what
actually fixed that problem.

BUG=chromium-os:30650
TEST=use requests to post a crash dump to the dev server 30 times in a row.
TEST=gmerge chromeos-chrome

Change-Id: I81097a6dfa3f6bf2b9f9599e803936b5e48f89ee
Reviewed-on: https://gerrit.chromium.org/gerrit/22433
Commit-Ready: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>