Revert "[dev-util] Remove server.socket_timeout from dev server"
You're change is breaking an au_test_harness test that tests updates in which the network stream 'hangs' for a bit.

http://chromegw/i/chromeos/builders/x86-alex32%20canary/builds/86

I believe that the test (which I wrote) is incorrectly hanging the following request in stateful_update (line 65), and your timeout change causes that request to fail.

  eval "wget -qS -T 300 -O - \"${STATEFUL_UPDATE_URL}\"" |
      tar --ignore-command-error --overwrite --directory=${STATEFUL_DIR} -xz

I'm reverting for now. The right fix might be to remove the hanging update test totally, I suspect it's of marginal benefit.

This reverts commit e4cdfa2e89fd97bd5b36eb0348f73f2f7f1a9a70

Change-Id: Ic12f02df29dea96812dd002f274e27e8e689ae38
Reviewed-on: https://gerrit.chromium.org/gerrit/22520
Commit-Ready: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>