[dev-util] Re-land "Remove server.socket_timeout from dev server"
This is what's causing us to wait for ages when we post crash dumps
to the dev server for symbolication.  It was put there as a part of
an attempt to fix an issue gmerg'ing chrome, but we're pretty sure
that other timeouts added as a part of the same change are what
actually fixed that problem.

BUG=chromium-os:30650
TEST=use requests to post a crash dump to the dev server 30 times in a row.
TEST=gmerge chromeos-chrome
TEST=remote trybot for x86-alex-release

CQ-DEPEND=If48d5684e1516e3f3f8b475896540136987b537d

This reverts commit 7aa142e39287b5e716f281ab493e9f3a09b1850b.

Change-Id: I436b882c1ead3753c77b7893a0b0cf823aee7786
Reviewed-on: https://gerrit.chromium.org/gerrit/22788
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>