cros_bundle_firmware: Pass the bundle into write_firmware
We need to be able to access the bundle from write_firmware, so we
can configure the BL2 of the flasher that we create. Add this as an
additional parameter.

BUG=chromium-os:28394
TEST=manual:
cros_bundle_firmware -b daisy -w usb
cros_bundle_firmware -b daisy -w sd:.
See that it still produces a booting image in each case

Change-Id: Ib4270f33a5c4f9b22060bdee57194ec5738d3dfc
Reviewed-on: https://gerrit.chromium.org/gerrit/23404
Commit-Ready: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>