cros_bundle_firmware: Support selection of a skeleton file
This file is provided by coreboot, and is the basis for our firmware image.
All the filename to be specified so that it can be used in the firmware
packing process. Provide a default filename also, for devs.

BUG=chrome-os-partner:7969
TEST=manual
emerge-link chromeos-bootimage
See that it picks up the skeleton file correct and builds an image

Change-Id: Id3694559868785d1384ce2f2f9b96df81924a74f
Reviewed-on: https://gerrit.chromium.org/gerrit/24766
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>