cros_bundle_firmware: Support selection of a skeleton file
This file is provided by coreboot, and is the basis for our firmware image.
All the filename to be specified so that it can be used in the firmware
packing process. Provide a default filename also, for devs.

BUG=chrome-os-partner:7969
TEST=manual
emerge-link chromeos-bootimage
See that it picks up the skeleton filskee correct and builds an image

Change-Id: I02c75111b768f0b758d285c76383cd5afe903c9c
Reviewed-on: https://gerrit.chromium.org/gerrit/24766
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/25299
Commit-Ready: Simon Glass <sjg@chromium.org>