cros_bundle_firmware: Handle ARCH_CPU_DTS for upstream U-Boot
This is a temporary hack to support upstream U-Boot. Since it does not
have the benefit of the dtc -i option, it uses the C preprocessor to find
its include files. That doesn't work for us since we have .dtsi files in
the cros/ subdirectory also. But we still want to be able to flash
an upstream U-Boot using cros_bundle_firmware, so must add this support.

BUG=chromium-os:31621
TEST=emerge-tegra2_seaboard chromeos-bootimage

Change-Id: I0af8036862d1a5a1db840f0db52de2ca46a7c36f
Reviewed-on: https://gerrit.chromium.org/gerrit/24763
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>